--- backend/src/main/java/com/shfh/pdmp/demo/curd/entities/UserEntity.java
@@ -1,34 +1,0 @@
-package com.shfh.pdmp.demo.curd.entities;
-
-public class UserEntity {
-
-    private String id;
-
-    private String username;
-
-    private int age;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-}
