--- backend/src/main/java/com/shfh/pdmp/demo/curd/service/IDemoService.java
@@ -1,7 +1,0 @@
-package com.shfh.pdmp.demo.curd.service;
-
-/**
- * Created by wpcao on 2018/7/14 16:54
- */
-public interface IDemoService{
-}
