--- backend/src/main/java/config/spring/spring-context.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-	   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	   xmlns:context="http://www.springframework.org/schema/context"
-	   xmlns:aop="http://www.springframework.org/schema/aop"
-	   xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-4.1.xsd
-		http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-4.1.xsd
-		http://www.springframework.org/schema/aop http://www.springframework.org/schema/aop/spring-aop.xsd"
-	   default-lazy-init="false">
-
-	<description>Spring公共配置</description>
-
-	<!-- 使用annotation 自动注册bean,并检查@Required,@Autowired的属性已被注入 -->
-	<context:component-scan base-package="com.shfh" />
-
-	<!-- 开启@AspectJ AOP代理 -->
-	<aop:aspectj-autoproxy proxy-target-class="true" />
-
-
-	<!--导入其他配置项-->
-	<import resource="spring-jdbc.xml"/>
-	<import resource="../**/*-spring.xml"/>
-</beans>
