--- backend/src/main/java/config/spring/spring-jdbc.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans" 
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
-
-	<!--config/xxx/xxx-datasource.xml-->
-	<import resource="../*/*-datasource.xml"/>
-
-</beans>
