--- backend/src/main/java/config/spring/spring-mvc.xml
@@ -1,82 +1,0 @@
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:context="http://www.springframework.org/schema/context"
-       xmlns:mvc="http://www.springframework.org/schema/mvc"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-4.3.xsd
-				http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-4.3.xsd
-				http://www.springframework.org/schema/mvc http://www.springframework.org/schema/mvc/spring-mvc-4.3.xsd">
-
-
-    <!-- 启动注解驱动的Spring MVC功能 -->
-    <mvc:annotation-driven/>
-
-    <!-- 自动扫描且只扫描@Controller -->
-    <!--<context:component-scan base-package="com.centit" use-default-filters="false">-->
-        <!--<context:include-filter type="annotation" expression="org.springframework.stereotype.Controller"/>-->
-    <!--</context:component-scan>-->
-    <context:component-scan base-package="com.shfh" use-default-filters="false">
-        <context:include-filter type="annotation" expression="org.springframework.stereotype.Controller"/>
-    </context:component-scan>
-
-    <mvc:annotation-driven content-negotiation-manager="contentNegotiationManager">
-        <mvc:message-converters register-defaults="true">
-            <bean class="org.springframework.http.converter.ByteArrayHttpMessageConverter"/>
-            <!-- 将StringHttpMessageConverter的默认编码设为UTF-8 -->
-            <bean class="org.springframework.http.converter.StringHttpMessageConverter">
-                <constructor-arg value="UTF-8"/>
-            </bean>
-            <!-- 将Jackson2HttpMessageConverter的默认格式化输出设为true -->
-            <bean
-                    class="org.springframework.http.converter.json.MappingJackson2HttpMessageConverter">
-                <property name="prettyPrint" value="true"/>
-            </bean>
-        </mvc:message-converters>
-    </mvc:annotation-driven>
-
-
-    <!-- 处理静态文件请求，对静态文件不进行 Dispatcher -->
-    <mvc:default-servlet-handler/>
-
-    <!-- REST中根据URL后缀自动判定Content-Type及相应的View -->
-    <bean id="contentNegotiationManager"
-          class="org.springframework.web.accept.ContentNegotiationManagerFactoryBean">
-        <property name="ignoreAcceptHeader" value="true"/>
-        <property name="defaultContentType" value="application/json"/>
-        <property name="mediaTypes">
-            <value>
-                json=application/json
-                xml=application/xml
-            </value>
-        </property>
-    </bean>
-
-
-    <!--跨域规则，仅支持/api/开头的请求跨域-->
-    <mvc:cors>
-        <mvc:mapping path="/api/**"
-                    allowed-origins="*"
-                    allowed-methods="GET,POST,OPTIONS,DELETE,PUT"
-                    allowed-headers="*"
-                    allow-credentials="true"/>
-    </mvc:cors>
-
-
-    <!-- 文件上传  -->
-    <bean id="multipartResolver"
-          class="org.springframework.web.multipart.commons.CommonsMultipartResolver">
-        <!-- 上传文件大小上限，单位为字节（10MB） -->
-        <property name="maxUploadSize" value="10485760"/>
-        <!-- 请求的编码格式，必须和jSP的pageEncoding属性一致，以便正确读取表单的内容，默认为ISO-8859-1 -->
-        <property name="defaultEncoding" value="UTF-8"/>
-    </bean>
-
-
-
-    <!--jsp视图解析器-->
-    <bean id="jspViewResolver" class="org.springframework.web.servlet.view.InternalResourceViewResolver">
-        <property name="viewClass" value="org.springframework.web.servlet.view.JstlView"/>
-        <property name="prefix" value="/WEB-INF/jsp/"/>
-        <property name="suffix" value=".jsp"/>
-        <property name="order" value="10"/>
-    </bean>
-</beans>
