--- backend/src/main/resources/index.jsp
@@ -1,11 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
-<% String contextPath = request.getContextPath();%>
-<html>
-<head>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-</head>
-<body>
-
-    Hello
-</body>
-</html>
