+++ src/main/java/com/fh/pdmp/base/ResponseConstant.java
@@ -1,0 +1,25 @@
+package com.fh.pdmp.base;
+
+public class ResponseConstant {
+	/**
+	 * 返回状态码
+	 */
+	public static final int CODE_SUCCESS = 0;
+	public static final int CODE_ERROR = -1;
+
+	/**
+	 * 返回状态信息
+	 */
+	public static final String MSG_SUCCESS = "操作成功";
+	public static final String MSG_ERROR = "操作失败";
+
+	/**
+	 * 默认分页页码
+	 */
+	public static final int DEFAULT_PAGE_NO = 1;
+
+	/**
+	 * 默认分页大小
+	 */
+	public static final int DEFAULT_PAGE_SIZE = 20;
+}
