+++ src/main/java/com/fh/pdmp/demo/dao/IDemoDao.java
@@ -1,0 +1,4 @@
+package com.fh.pdmp.demo.dao;
+
+public interface IDemoDao {
+}
