+++ src/main/java/com/fh/pdmp/demo/dao/impl/DemoDaoImpl.java
@@ -1,0 +1,4 @@
+package com.fh.pdmp.demo.dao.impl;
+
+public class DemoDaoImpl {
+}
