+++ src/main/java/com/fh/pdmp/demo/service/IDemoService.java
@@ -1,0 +1,4 @@
+package com.fh.pdmp.demo.service;
+
+public class IDemoService {
+}
