+++ src/main/java/com/fh/pdmp/demo/service/impl/DemoServiceImpl.java
@@ -1,0 +1,9 @@
+package com.fh.pdmp.demo.service.impl;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class DemoServiceImpl {
+
+
+}
