+++ src/main/resources/config/applicationContext.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<beans xmlns="http://www.springframework.org/schema/beans"
+	   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	   xmlns:context="http://www.springframework.org/schema/context"
+	   xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-4.1.xsd
+		http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-4.1.xsd">
+
+	<description>Spring公共配置</description>
+
+	<!-- 使用annotation 自动注册bean -->
+	<context:component-scan base-package="com.fh.pdmp"/>
+
+	<!--导入其他配置项-->
+	<import resource="spring-servlet.xml"/>
+</beans>
