+++ src/main/resources/config/spring-servlet.xml
@@ -1,0 +1,34 @@
+<beans xmlns="http://www.springframework.org/schema/beans"
+	   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	   xmlns:context="http://www.springframework.org/schema/context"
+	   xmlns:mvc="http://www.springframework.org/schema/mvc"
+	   xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-4.3.xsd
+				http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-4.3.xsd
+				http://www.springframework.org/schema/mvc http://www.springframework.org/schema/mvc/spring-mvc-4.3.xsd">
+
+
+	<!-- 自动扫描 -->
+	<context:component-scan base-package="com.fh.pdmp.demo.controller"/>
+
+	<!-- 启动注解驱动的Spring MVC功能（设置resp返回值编码为utf-8） -->
+	<mvc:annotation-driven>
+		<mvc:message-converters register-defaults="true">
+			<bean class="org.springframework.http.converter.StringHttpMessageConverter">
+				<constructor-arg value="UTF-8" />
+			</bean>
+		</mvc:message-converters>
+	</mvc:annotation-driven>
+
+	<!-- 处理静态文件请求，对静态文件不进行 Dispatcher -->
+	<mvc:default-servlet-handler/>
+
+	<!-- configure the InternalResourceViewResolver -->
+	<bean class="org.springframework.web.servlet.view.InternalResourceViewResolver"
+		  id="internalResourceViewResolver">
+		<!-- 前缀 -->
+		<property name="prefix" value="/WEB-INF/"/>
+		<!-- 后缀 -->
+		<property name="suffix" value=".html"/>
+	</bean>
+
+</beans>
