+++ src/main/resources/index.html
@@ -1,0 +1,1 @@
+Hello
